+++ src/main/java/envios/envio/RetiroEnSucursal.java
@@ -2,7 +2,7 @@ package envios.envio;
 
 public class RetiroEnSucursal implements MetodoDeEnvio {
 
-    public float costoDeEnvio(float precioTotalPedido, float peso, String direccionEnvio) {
+    public double costoDeEnvio(double precioTotalPedido, double peso, String direccionEnvio) {
         return 0;
     }
 }
